+++ content/__tests__/rating-menu-coordinator.test.ts
