+++ content/auto-reset.ts
