+++ content/rating-menu-coordinator.ts
