+++ content/rating-menu.ts
